Refactor outage counter availability logic
